windows fix
git-svn-id: https://deskvox.svn.sourceforge.net/svnroot/deskvox/trunk@1578 04231f92-3938-0410-9931-e931ac552b4f

+++ virvo/virvo/vvsocketio.cpp
@@ -775,7 +775,7 @@ vvSocket::ErrorType vvSocketIO::putFileName(const char* fn)
 
   const size_t len = fn ? strlen(fn) : 0;
   buffer = new uchar[4 + len];
-  vvToolshed::write32(&buffer[0], len);
+  vvToolshed::write32(&buffer[0], (uint32_t)len);
 
   for (size_t i=0; i<len; ++i)
   {
